refactor: Re-structured codebase
