+++ src/index.ts
@@ -1,8 +1,8 @@
-export * from './create-route';
-export * from './new-create-history-router';
-export * from './create-router-controls';
+export * from './methods/create-route';
+export * from './methods/new-create-history-router';
+export * from './methods/create-router-controls';
 export * from './utils/build-path';
-export * from './utils/is-route';
-export * from './utils/redirect';
-export * from './utils/chain-route';
+export * from './methods/is-route';
+export * from './methods/redirect';
+export * from './methods/chain-route';
 export * from './types';
