+++ src/methods/is-route.ts
